+++ naturtag/utils/image_glob.py
@@ -17,14 +17,22 @@ def get_valid_image_paths(
     paths_or_uris: Iterable[PathOrStr],
     recursive: bool = False,
     include_sidecars: bool = False,
+    create_sidecars: bool = False,
 ) -> set[Path]:
     """
     Get all images of supported filetypes from one or more dirs and/or image paths, including URIs.
+
+    Notes on sidecar files:
+
+    * Directly passing a path to a sidecar file is allowed
+    * When passing a path to a file that is not writeable (e.g., RAW images), a sidecar file will be
+      created or updated for it
 
     Args:
         paths: Paths or file URIs to images and/or image directories
         recursive: Recursively get images from subdirectories
         include_sidecars: Allow loading a sidecar file without an associated image
+        create_sidecars: Create a new sidecar file if a non-writeable file path is provided
 
     Returns:
          Combined list of image file paths
@@ -38,11 +46,19 @@ def get_valid_image_paths(
     for path in paths_or_uris:
         if not path:
             continue
+
         path = uri_to_path(path)
+        sidecar_path = get_sidecar_path(path)
         if path.is_dir():
             image_paths.extend(get_images_from_dir(path, recursive=recursive))
         elif is_image_path(path, include_sidecars=include_sidecars):
             image_paths.append(path)
+        elif include_sidecars and sidecar_path.is_file():
+            logger.debug(f'{path} is not writable; using existing sidecar: {sidecar_path}')
+            image_paths.append(sidecar_path)
+        elif path.is_file() and create_sidecars:
+            logger.debug(f'{path} is not writable; creating sidecar: {sidecar_path}')
+            image_paths.append(sidecar_path)
         else:
             logger.warning(f'Not a valid path: {path}')
 
@@ -83,6 +99,12 @@ def glob_paths(path_patterns: Iterable[PathOrStr]) -> list[Path]:
             glob(str(pattern), recursive=True) for pattern in path_patterns
         )
     ]
+
+
+def get_sidecar_path(path: Path) -> Path:
+    default_path = path.with_suffix('.xmp')
+    alt_path = path.with_suffix(f'{path.suffix}.xmp')
+    return alt_path if alt_path.is_file() else default_path
 
 
 def is_image_path(path: Path, include_sidecars: bool = False) -> bool: